Extract normalizeEvaluateAsCommentArgs to evaluate-utils module
Move the pure argument normalization logic out of evaluate.ts into
evaluate-utils.ts so that unit tests can import it without pulling
in the vscode dependency.
